save rgsc data, test smrnov in prod

+++ TavisApi/Interfaces/IBcmService.cs
@@ -6,5 +6,5 @@ public interface IBcmService
 {
   List<Player> GetPlayers();
   DateTime? GetContestStartDate();
-  double? CalcBcmValue(double? ratio, double? estimate);
+  int? CalcBcmValue(double? ratio, double? estimate);
 }
